requirements.txt
chardet==5.2.0
gitdb==4.0.11
GitPython==3.1.41
python-tlsh==4.5.0
setuptools==69.0.3
smmap==5.0.1
wheel==0.43.0
